+++ skcriteria/cmp/ranks_cmp.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # License: BSD-3 (https://tldrlegal.com/license/bsd-3-clause-license-(revised))
 # Copyright (c) 2016-2021, Cabral, Juan; Luczywo, Nadia
-# Copyright (c) 2022, 2023, 2024 QuatroPe
+# Copyright (c) 2022-2025 QuatroPe
 # All rights reserved.
 
 # =============================================================================
@@ -88,11 +88,13 @@ class RanksComparator(Sequence, DiffEqualityMixin):
 
     def __init__(self, ranks):
         ranks = list(ranks)
-        self._validate_ranks(ranks)
         self._ranks = ranks
+        self._validate_ranks()
 
     # INTERNALS ===============================================================
-    def _validate_ranks(self, ranks):
+    def _validate_ranks(self):
+        ranks = self._ranks
+
         if len(ranks) <= 1:
             raise ValueError("Please provide more than one ranking")
 
@@ -227,6 +229,8 @@ class RanksComparator(Sequence, DiffEqualityMixin):
         df.columns.name = "Method"
 
         return df
+
+    # STATISTICALS ============================================================
 
     def corr(self, *, untied=False, **kwargs):
         """Compute pairwise correlation of rankings, excluding NA/null values.
@@ -370,6 +374,41 @@ class RanksComparator(Sequence, DiffEqualityMixin):
             dis_mtx, columns=df.index.copy(), index=df.index.copy()
         )
         return dis_df
+
+    def extra_get(self, key, default=None):
+        """Retrieve a specific key from each rank, returning a \
+        dictionary of results.
+
+        This method iterates through all ranks and attempts to get the value
+        associated with the specified key. If the key is not found in a rank,
+        the default value is used.
+
+        Parameters
+        ----------
+        key : hashable
+            The key to look up in each rank.
+        default : any, optional
+            The value to return if the key is not found in a rank.
+            Defaults to None.
+
+        Returns
+        -------
+        dict
+            A dictionary where each key is the name of a rank, and each value
+            is the result of calling `get(key, default)` on that rank.
+
+        Notes
+        -----
+        The returned dictionary will have an entry for every rank, even if the
+        key was not found and the default value was used.
+
+        """
+        return {
+            rank_name: rank.extra_.get(key, default)
+            for rank_name, rank in self._ranks
+        }
+
+    eget = extra_get  # shortcut
 
     # ACCESSORS (YES, WE USE CACHED PROPERTIES IS THE EASIEST WAY) ============