+++ skcriteria/utils/bunch.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # License: BSD-3 (https://tldrlegal.com/license/bsd-3-clause-license-(revised))
 # Copyright (c) 2016-2021, Cabral, Juan; Luczywo, Nadia
-# Copyright (c) 2022, 2023, 2024 QuatroPe
+# Copyright (c) 2022-2025 QuatroPe
 # All rights reserved.
 
 # =============================================================================
@@ -46,17 +46,14 @@ class Bunch(Mapping):
     >>> b.a = 3
     >>> b['a']
     3
-    >>> b.c = 6
-    >>> b['c']
-    6
 
     """
 
     def __init__(self, name, data):
         if not isinstance(data, Mapping):
             raise TypeError("Data must be some kind of mapping")
-        self._name = str(name)
-        self._data = data
+        super().__setattr__("_name", str(name))
+        super().__setattr__("_data", data)
 
     def __getitem__(self, k):
         """x.__getitem__(y) <==> x[y]."""
@@ -68,6 +65,10 @@ class Bunch(Mapping):
             return self._data[a]
         except KeyError:
             raise AttributeError(a)
+
+    def __setattr__(self, a, v):
+        """x.__setattr__(a, v) <==> x.a = v."""
+        raise AttributeError(f"Bunch {self._name!r} is read-only")
 
     def __copy__(self):
         """x.__copy__() <==> copy.copy(x)."""
@@ -87,7 +88,7 @@ class Bunch(Mapping):
         memo[id(self)] = clone
 
         # now we copy the data
-        clone._data = copy.deepcopy(self._data, memo)
+        super(cls, clone).__setattr__("_data", copy.deepcopy(self._data, memo))
 
         return clone
 
@@ -107,3 +108,11 @@ class Bunch(Mapping):
     def __dir__(self):
         """x.__dir__() <==> dir(x)."""
         return super().__dir__() + list(self._data)
+
+    def __setstate__(self, state):
+        """Needed for multiprocessing environment."""
+        self.__dict__.update(state)
+
+    def get(self, key, default=None):
+        """Get item from bunch."""
+        return self._data.get(key, default)