+++ tests/core/test_data.py
@@ -2,16 +2,14 @@
 # -*- coding: utf-8 -*-
 # License: BSD-3 (https://tldrlegal.com/license/bsd-3-clause-license-(revised))
 # Copyright (c) 2016-2021, Cabral, Juan; Luczywo, Nadia
-# Copyright (c) 2022, 2023, 2024 QuatroPe
+# Copyright (c) 2022-2025 QuatroPe
 # All rights reserved.
 
 # =============================================================================
 # DOCS
 # =============================================================================
 
-"""test for skcriteria.core.data
-
-"""
+"""test for skcriteria.core.data"""
 
 
 # =============================================================================
@@ -300,6 +298,26 @@ def test_DecisionMatrix_copy(data_values):
     assert dm is not copy
     assert dm.equals(copy)
 
+    with pytest.deprecated_call():
+        dm.copy(**dm.to_dict())
+
+
+def test_DecisionMatrix_replace(data_values):
+    mtx, objectives, weights, alternatives, criteria = data_values(seed=42)
+
+    dm = data.mkdm(
+        matrix=mtx,
+        objectives=objectives,
+        weights=weights,
+        alternatives=alternatives,
+        criteria=criteria,
+    )
+    copy = dm.replace(weights=dm.weights + 1)
+
+    assert dm is not copy
+    assert not dm.equals(copy)
+    pd.testing.assert_series_equal(dm.weights, copy.weights - 1)
+
 
 def test_DecisionMatrix_to_dataframe(data_values):
     mtx, objectives, weights, alternatives, criteria = data_values(seed=42)
@@ -346,6 +364,32 @@ def test_DecisionMatrix_to_dict(data_values):
 
     cmp = {k: (np.all(result[k] == expected[k])) for k in result.keys()}
     assert np.all(cmp.values())
+
+
+def test_DecisionMatrix_to_latex(data_values):
+    mtx, objectives, weights, alternatives, criteria = data_values(seed=42)
+
+    dm = data.mkdm(
+        matrix=mtx,
+        objectives=objectives,
+        weights=weights,
+        alternatives=alternatives,
+        criteria=criteria,
+    )
+
+    latex = dm.to_latex()
+
+    # create the expected table
+    df = dm.to_dataframe()
+    df.columns = [rf"\textbf{{{col}}}" for col in df.columns]
+
+    expected = df.to_latex(bold_rows=True)
+
+    expected_lines = expected.splitlines()
+    expected_lines.insert(6, r"\midrule")
+    expected = "\n".join(expected_lines)
+
+    assert latex == expected
 
 
 def test_DecisionMatrix_describe(data_values):